# financeApp-backend-feature-admin-group-management/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finance Backend API",
        "description": "RESTful API backend for dual-version finance management application supporting Admin and User versions of Flutter mobile app",
        "contact": {
            "email": "support@financeapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/group": {
            "get": {
                "tags": [
                    "Admin Group Management"
                ],
                "summary": "Get admin's group information",
                "description": "Returns the authenticated admin's group information including group code",
                "operationId": "24dc631af3c24c439c0ea7a0ff94f5ff",
                "responses": {
                    "200": {
                        "description": "Group information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Group information retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "group": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "admin_user_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "group_code": {
                                                            "type": "string",
                                                            "example": "123456"
                                                        },
                                                        "group_name": {
                                                            "type": "string",
                                                            "example": "Acme Corp - John Doe"
                                                        },
                                                        "is_active": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin does not have a group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admin does not have a group."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/group/regenerate": {
            "post": {
                "tags": [
                    "Admin Group Management"
                ],
                "summary": "Regenerate admin's group code",
                "description": "Generates a new unique group code for the authenticated admin",
                "operationId": "c2f0cc39e4acf7672160c9064f02cd71",
                "responses": {
                    "200": {
                        "description": "Group code regenerated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Group code regenerated successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "group": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "admin_user_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "group_code": {
                                                            "type": "string",
                                                            "example": "654321"
                                                        },
                                                        "group_name": {
                                                            "type": "string",
                                                            "example": "Acme Corp - John Doe"
                                                        },
                                                        "is_active": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin does not have a group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to regenerate group code."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}